+++ src/Classifiers/ClassificationTree.php
@@ -65,14 +65,16 @@ class ClassificationTree extends CART implements Estimator, Learner, Probabilist
      * @param int $maxLeafSize
      * @param float $minPurityIncrease
      * @param int|null $maxFeatures
+     * @param int|null $maxBins
      */
     public function __construct(
         int $maxHeight = PHP_INT_MAX,
         int $maxLeafSize = 3,
         float $minPurityIncrease = 1e-7,
-        ?int $maxFeatures = null
+        ?int $maxFeatures = null,
+        ?int $maxBins = null
     ) {
-        parent::__construct($maxHeight, $maxLeafSize, $minPurityIncrease, $maxFeatures);
+        parent::__construct($maxHeight, $maxLeafSize, $minPurityIncrease, $maxFeatures, $maxBins);
     }
 
     /**
@@ -116,6 +118,7 @@ class ClassificationTree extends CART implements Estimator, Learner, Probabilist
             'max leaf size' => $this->maxLeafSize,
             'min purity increase' => $this->minPurityIncrease,
             'max features' => $this->maxFeatures,
+            'max bins' => $this->maxBins,
         ];
     }